pyproject.toml
[project]
name = "spacer_matching_bench"
dynamic = ["version"]
authors = [
    { name = "Uri Neri", email = "uneri@lbl.gov" },
    { name = "Rick Beeloo", email = "biobeeloo@gmail.com" },
    { name = "Simon Roux", email = "sroux@lbl.gov" },
    { name = "Antonio Camargo", email = "antoniop.camargo@lbl.gov" },
    { name = "Brian Bushnell", email = "bbushnell@lbl.gov" },
]
description = "Benchmarking protospacer identification tools"
readme = "README.md"
requires-python = ">=3.9"
license = { file = "LICENSE" }
dependencies = [
    "needletail>=0.7.0,<0.8",
    "pyfastx>=2.2.0,<3",
    "pysam>=0.23.0,<0.24",
    "hyperfine>=0.1.3,<0.2",
    "ipython>=8.18.1,<9",
    "tqdm>=4.67.1,<5",
    "parasail>=1.3.4,<2",
    "altair>=5.5.0,<6",
    "ipykernel>=6.29.5,<7",
    "pixi-kernel>=0.6.3,<0.7",
    "jupyter>=1.1.1,<2",
    "jupyter-core>=5.7.2,<6",
    "executing>=2.2.0,<3",
    "upsetplot>=0.9.0,<0.10",
    "matplotlib>=3.9.4,<4",
    "vl-convert-python>=1.7.0,<2",
    "svgutils>=0.3.4,<0.4",
    "itables>=2.3.0,<3",
    "seaborn>=0.13.2,<0.14",
    "vegafusion>=2.0.2,<3",
    "cmap>=0.6.0,<0.7",
    "pypdf2>=3.0.1,<4",
    "charset-normalizer>=3.4.2,<4", "polars>=1.34.0,<2",
]

[project.urls]
Source = "https://code.jgi.doe.gov/spacersdb/spacer_matching_bench"
Documentation = "https://code.jgi.doe.gov/spacersdb/spacer_matching_bench/docs/"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
packages = ["bench"] #this is the package name

[project.scripts]
spacer_bencher = "bench.bench:main" # this is the CLI entry point

[dependency-groups]
build = ["hatch >=1.14.0,<2", "pip >=25.0.1,<26"]
format = ["ruff >=0.11.3,<0.12"]
publish = ["twine >=6.1.0,<7"]

[tool.hatch.version]
path = "src/bench/__init__.py"

[tool.hatch.build.targets.wheel]
packages = ["src/bench"]                #this is the package name (again)
package-dir = { "bench" = "src/bench" } #this is the package directory

[tool.pixi.workspace]
channels = ["conda-forge", "bioconda"]
platforms = ["linux-64"]

[tool.pixi.pypi-dependencies]
spacer_matching_bench = { path = ".", editable = true }

[tool.pixi.dependencies]
pigz = ">=2.8,<3"
seqkit = ">=2.10.0,<3"
pip = ">=25.1.1,<26"
rust = ">=1.86.0,<1.87"
hatch = ">=1.14.1,<2"
maturin = ">=1.8.3,<2"
quarto = ">=1.6.43,<2"
librsvg = ">=2.58.4,<3"

[tool.pixi.feature.py39.dependencies]
python = "~=3.9.0"

[tool.pixi.environments]
default = { features = ["build", "format", "publish", "py39", "render"] }

[tool.pixi.feature.build.tasks]
update-python-version = "hatch version micro && hatch build --clean "
build-python = "hatch build --clean && pip install -e ."
build-rust = "cd src/rust_simulator && maturin develop --release && cd ../../"
build-all = { depends-on = ["build-rust", "build-python"] }

[tool.pixi.feature.render.tasks]
render = "bash ./draft/build_docs.sh"

[tool.ruff]
lint.ignore = [
    "E402", # Module level import not at top of file
    "F403", # import * / *` used; unable to detect undefined names
    "F405", # name may be undefined, or defined from star imports: module
]
